Revert "Refactor proactor pool, add frontier-carrying signals, and fix shared infra. (#23804)"
This reverts commit 98078db0603e220419721545c978ff0b3b970dcf.
