+++ build_tools/sanitizer/lsan_suppressions_vulkan.txt
@@ -1,10 +1,7 @@
-# LSAN suppressions for Vulkan driver leaks.
-# Mesa/RADV allocates per-thread state via pthread_once that is intentionally
-# never freed (driver thread-local caches). These are driver-internal
-# allocations, not IREE bugs.
+# LSAN suppressions for Vulkan ICD leaks
+# Vulkan ICDs allocate thread-local state via pthread_once callbacks on
+# internally-spawned threads. This state is never freed. These leaks are
+# in external GPU driver code (reported as <unknown module> by ASAN since
+# the symbolizer cannot resolve ICD internals).
 
-leak:libvulkan_radeon.so
-leak:libvulkan_lvp.so
-# Fallback: when ASAN can't resolve the shared library name, match the
-# pthread_once call that triggers the per-thread initialization.
 leak:__pthread_once_slow
