+++ runtime/src/iree/hal/drivers/hip/hip_device.c
@@ -1809,8 +1809,7 @@ static iree_status_t iree_hal_hip_device_queue_alloca(
         out_buffer);
   }
   if (iree_status_is_ok(status)) {
-    status = iree_hal_semaphore_list_signal(signal_semaphore_list,
-                                            /*frontier=*/NULL);
+    status = iree_hal_semaphore_list_signal(signal_semaphore_list);
   }
   if (iree_status_is_ok(status)) {
     iree_hal_hip_device_advance_frontier(device);
@@ -1917,8 +1916,7 @@ static iree_status_t iree_hal_hip_device_queue_dealloca(
         device->devices[device_ordinal].hip_dispatch_stream, buffer);
   }
   if (iree_status_is_ok(status)) {
-    status = iree_hal_semaphore_list_signal(signal_semaphore_list,
-                                            /*frontier=*/NULL);
+    status = iree_hal_semaphore_list_signal(signal_semaphore_list);
   }
   if (iree_status_is_ok(status)) {
     iree_hal_hip_device_advance_frontier(device);
